(lpf) midi input
To figure how to map a frequency to a lpf lowness....

First, convince yourself that absolute frequency doesn't matter.
All we care to achieve is proper frequency ratio between two midi notes given to an lpf.
This is particularly useful to avoid philosophy on what the frequency of an lpf lowness is,
and to make an easier analogy between dft bins and frequency.

As for finding the quadratic equation, understand this is, as far as I'm concerned,
just a regression that works well. Might go deeper than that, but I didn't delve.

The parabola is regressed on a set of points obtained with the following....

```
from numpy.fft import fft
import matplotlib.pyplot as plt

#impulse response of lpf
def ir(lowness, n=32): return [(1-lowness)*lowness**i for i in range(n)]

def indexOfHalf(f):
 for i, x in enumerate(f):
  if abs(x) < 0.5: return i

#for lownesses 0.25 to 1.0,
#get frequency response of impulse response
#find at which frequency bin amplitude goes below half
plt.plot([indexOfHalf(fft(ir(i/1000, 512))) for i in range(250, 1000)]); plt.show()
```

Given this plot, we desire a parabola that does the reverse: lowness given frequency.
We also expect c to be 1. It'd be smarter to express this in the regressor,
but I just added a bunch of (0, 1) points.

You can demo this with the sonic system by making noise with the sonic.