remove looper
I like the idea but I'm realizing there's too much distribution of implementation.
This should be entirely a client to the audio system.
As written, all functionality is implemented in the process containing the audio system.
It's split into "backend" stuff that somewhat fits here, and "frontend" stuff that definitely doesn't.
The backend stuff fits equally well in a client context.
The frontend stuff is configuration, that a naive softboard interface then interacts with.