remove editor
This shold be rebuilt as a web app.
Some day it might make sense to revive similar as a desktop app.
However, the assumptions it makes about the system and network and commander components are going away.
So I'd rather just discard it for now, get the inner architecture right, then build out again.