remove network stuff
Initially I hoped to cleanly cut this out of `System`.
I forgot that I set up network components to be dependent on this.
Oh well, let's get rid of network components.
We'll rebuild with web stuff.

qweboard and softboard go away with network
pipe is no longer needed if qweboard is gone