remove replay system
I'm not sure what the future of the filei and fileo components is.

Their original intent was to enable recording and replaying entire sessions constructing and exploring an audio system.
This probably made more sense when commander components were favored for interacting with a running system.
Now it seems to make more sense to simply record all traffic coming into dlalRequest.

I think it makes sense to keep these components around.
I'm not entirely sure what they'll do, but that's not the point.

I think the replay system was built around the original intent.
Now it's superceded by something like the ambient system.