(web audio) don't forget to import sleep, naming convention fix, smoother sleep
Right now if requests from client to server get choked up,
the server will hit its buffer limit and drop samples.
Similarly, if client refreshes, audio will stutter after a server buffer's worth of audio.

The eventual fix to this is to have server offer messages to the websocket,
rather than having a client request.
This kind of thing is important if there are to be multiple clients -- right now only one client will work.

The challenge -- and it may not be a big one -- is how to speak with a websocket from the audio callback without delaying.
Perhaps a separate thread could be notified by the audio callback that it is time to do some websocket work.