fix a crash introduced in HEAD~4 (oracle)
"View command borrows body instead of moving" seems to have aggravated a problem with vec_u8.
This way of doing things avoids a transmute.
Not sure how to explain the crash, but it's gone now.
The crash happens whenever the midi command is called.
The crash can be reproduced by playing a note on the webboard.