(buf) zero
will figure how to accumulate-over-time when a use case comes up
mainly seems useful as a delay (echo),
but a dedicated component seems more appropriate for that

might be some intersection when buf learns to record
I didn't like how previously controlling recording depended on changing connections,
I think it makes more sense to start a recording on midi note on,
and end it (and learns its duration) on midi note off.
Begs the question how to switch between record and replay modes though.
Maybe the buf could discern a MIDI channel for playing (likely a liner) and a MIDI channel for recording (likely a midi).

recording also brings looping into question,
but I think looping fits better as MIDI,
because it's much easier to do some features with MIDI:
- loop detection
- remote collaborative editing