thoughts
I don't like the markov approach.
To be clear, it's come to be somewhat synonymous with the bucketing idea,
which I think is along the right lines.
But I don't like the loss of control when synthesizing with a markov model.
And I've already had to babysit it to not entirely go off the rails
(or stay on insane rails, really) and it's still not fully sane.

We already have a way to go from reduced params to full params.
What we lack is a way to generate reduced params.
The markov model attempts to solve both at once,
but it does poorly at generating reduced params.
Thinking exclusively of a markov model's ability to generate reduced params,
it's in the uncanny valley between something simple and predictable,
and a well-tuned neural net.

I'm hoping applying human intellect to a visualization of features can yield a simple generator.